Changed attack to hitbox detection

+++ Vampire Roguelike/Assets/Scripts/Enemy.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class Enemy : MonoBehaviour
 {
-    public int health;
+    public float health;
     public int damage;
     public float speed;
     public SpriteRenderer sprite;
@@ -31,7 +31,7 @@ public class Enemy : MonoBehaviour
         }
     }
 
-    public void takeDamage(int damage)
+    public void takeDamage(float damage)
     {
         Instantiate(bloodParticle, transform.position, Quaternion.identity);
         health -= damage;
